New function addMapLegend. addMapLegend requires the fields package, so rworldmap now suggests fields.
addMapLegend adds legends to plots created by mapCountryData and mapGridAscii. Can be used after map functions, or if addLegend=TRUE, and you have the fields package, it is now the default legend. If you do not have fields, addMapLegend is not used, and the old style legend remains. 

Also minor doc tweak, previously topo was an undocumented option for the colourPalette parameter.

git-svn-id: svn://svn.r-forge.r-project.org/svnroot/rworldmap@10 edb9625f-4e0d-4859-8d74-9fd3b1da38cb